OCRModules-V4/requirements.txt
paddlepaddle==2.1.2
onnx==1.7.0
onnx-simplifier==0.2.10
onnxruntime==1.3.0
openvino-dev==2023.0.1
torch==1.6.0
torchvision==0.7.0
pyclipper==1.3.0